requirements.txt
# PSA Global Insights Dashboard - Requirements
# Python 3.11+

# ============================================
# FRONTEND - Streamlit & Visualization
# ============================================
streamlit==1.28.0
plotly==5.17.0
numpy==1.25.2
pandas==2.0.3

# ============================================
# BACKEND - AI/ML Services
# ============================================
openai==1.3.0
# Azure OpenAI client is included in openai package

# ============================================
# DATA LAYER - Database & APIs
# ============================================
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
pdfplumber==0.10.3
PyPDF2==3.0.1
tabula-py==2.8.2

# ============================================
# UTILITIES
# ============================================
# Date/Time handling
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Configuration management
pydantic==2.4.2
pydantic-settings==2.0.3

# ============================================
# TESTING (Optional)
# ============================================
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# ============================================
# DEPLOYMENT
# ============================================
gunicorn==21.2.0
uvicorn==0.23.2

# ============================================
# DEVELOPMENT (Optional)
# ============================================
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0

# ============================================
# OPTIONAL - Enhanced Features
# ============================================
# Caching
redis==5.0.0

# Task scheduling
schedule==1.2.0
APScheduler==3.10.4

# Performance monitoring
prometheus-client==0.18.0

# ============================================
# NOTES
# ============================================
# Install with: pip install -r requirements.txt
# For development: pip install -r requirements.txt -r requirements-dev.txt
